Move the default implementation into a separate package.
