+++ README.md
@@ -1,4 +1,4 @@
 [![Java CI](https://github.com/monarch-initiative/BioNERApi/workflows/Java%20CI%20with%20Maven/badge.svg)](https://github.com/monarch-initiative/BioNERApi/actions/workflows/maven.yml)
 
 # BioNERApi
-Bio Namd Entity Recognition API
+Bio Named Entity Recognition API
