+++ bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/Concept.java
@@ -3,6 +3,7 @@ package org.monarchinitiative.bionerapi.model;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.monarchinitiative.bionerapi.model.impl.ConceptDefault;
 import org.monarchinitiative.bionerapi.model.impl.TermIdSerializer;
 import org.monarchinitiative.phenol.ontology.data.Identified;
 import org.monarchinitiative.phenol.ontology.data.TermId;
