--- bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/RequestDefault.java
@@ -1,8 +1,0 @@
-package org.monarchinitiative.bionerapi.model;
-
-record RequestDefault(String payload) implements Request {
-    @Override
-    public String getPayload() {
-        return payload;
-    }
-}
