--- bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/ResponseDefault.java
@@ -1,8 +1,0 @@
-package org.monarchinitiative.bionerapi.model;
-
-record ResponseDefault(Result result) implements Response {
-    @Override
-    public Result getResult() {
-        return result;
-    }
-}
