+++ bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/impl/ConceptDefault.java
@@ -1,8 +1,9 @@
-package org.monarchinitiative.bionerapi.model;
+package org.monarchinitiative.bionerapi.model.impl;
 
+import org.monarchinitiative.bionerapi.model.Concept;
 import org.monarchinitiative.phenol.ontology.data.TermId;
 
-record ConceptDefault(TermId id, int start, int end, boolean excluded) implements Concept {
+public record ConceptDefault(TermId id, int start, int end, boolean excluded) implements Concept {
 
     @Override
     public TermId getId() {
