+++ bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/impl/RequestDefault.java
@@ -1,0 +1,10 @@
+package org.monarchinitiative.bionerapi.model.impl;
+
+import org.monarchinitiative.bionerapi.model.Request;
+
+public record RequestDefault(String payload) implements Request {
+    @Override
+    public String getPayload() {
+        return payload;
+    }
+}
