+++ bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/impl/ResponseDefault.java
@@ -1,0 +1,11 @@
+package org.monarchinitiative.bionerapi.model.impl;
+
+import org.monarchinitiative.bionerapi.model.Response;
+import org.monarchinitiative.bionerapi.model.Result;
+
+public record ResponseDefault(Result result) implements Response {
+    @Override
+    public Result getResult() {
+        return result;
+    }
+}
