+++ bioner-api-model/src/main/java/org/monarchinitiative/bionerapi/model/impl/ResultDefault.java
@@ -1,0 +1,18 @@
+package org.monarchinitiative.bionerapi.model.impl;
+
+import org.monarchinitiative.bionerapi.model.Concept;
+import org.monarchinitiative.bionerapi.model.Result;
+
+import java.util.Collection;
+
+public record ResultDefault(String payload, Collection<? extends Concept> concepts) implements Result {
+    @Override
+    public String payload() {
+        return payload;
+    }
+
+    @Override
+    public Collection<? extends Concept> getConcepts() {
+        return concepts;
+    }
+}
